feat(swiftui home): extract common footer onto HomeCardFooterView
