+++ PocketKit/Sources/PocketKit/Home/SwiftUI/Views/HomeCarouselView.swift
@@ -41,7 +41,15 @@ struct HomeCarouselView: View {
         VStack(alignment: .leading) {
             makeTopContent()
             Spacer()
-            makeFooter()
+            HomeCardFooterView(
+                model: model,
+                domain: currentItem?.bestDomain,
+                timeToRead: currentItem?.timeToRead,
+                isSaved: currentSavedItem != nil && currentSavedItem?.isArchived == false,
+                isArchived: currentSavedItem?.isArchived == true,
+                isFavorite: currentSavedItem?.isFavorite == true,
+                isSyndicated: currentItem?.isSyndicated == true
+            )
         }
         .padding()
         .background(Color(.ui.homeCellBackground))
@@ -85,100 +93,6 @@ private extension HomeCarouselView {
                 .clipped()
             Spacer()
         }
-    }
-    // TODO: SWIFTUI - Extract footer in a view since it's basically the same between hero and carousel.
-    /// Footer
-    func makeFooter() -> some View {
-        HStack(alignment: .bottom) {
-            makeFooterDescription()
-            Spacer()
-            HStack(alignment: .center) {
-                makeActionButton()
-                makeOverflowMenu()
-            }
-        }
-    }
-
-    /// Footer description
-    func makeFooterDescription() -> some View {
-        VStack(alignment: .leading, spacing: Constants.stackSpacing) {
-            if let domain = currentItem?.bestDomain {
-                makeDomain(domain)
-                    .style(model.domainStyle)
-                    .lineLimit(Constants.footerElementLineLimit)
-            }
-            if let timeToRead = currentItem?.timeToRead, timeToRead > 0 {
-                Text(model.timeToRead(timeToRead))
-                    .lineLimit(Constants.footerElementLineLimit)
-            }
-        }
-    }
-
-    func makeDomain(_ domain: String) -> Text {
-        if currentItem?.isSyndicated == true {
-            return Text(domain) + Text(" ") + Text(Image(systemName: "checkmark.seal"))
-        } else {
-            return Text(domain)
-        }
-    }
-
-    /// Action button: save/saved and/or favorite
-    @ViewBuilder
-    func makeActionButton() -> some View {
-        HStack(alignment: .bottom) {
-            if let favoriteAction = model.favoriteAction {
-                makeFavoriteButton(handler: favoriteAction.action)
-            }
-            makeSaveButton()
-        }
-    }
-
-    func makeFavoriteButton(handler: @escaping (() -> Void)) -> some View {
-        HomeActionButton(
-            isActive: currentSavedItem?.isFavorite == false,
-            activeImage: .favoriteFilled,
-            inactiveImage: .favorite,
-            highlightedColor: .branding.amber1,
-            activeColor: .branding.amber4,
-            inactiveColor: .ui.grey8
-        ) {
-            handler()
-        }
-        .accessibilityIdentifier("save-button")
-    }
-
-    func makeSaveButton() -> some View {
-        HomeActionButton(
-            isActive: currentSavedItem?.isArchived == false,
-            activeImage: .saved,
-            inactiveImage: .save,
-            activeTitle: Localization.Recommendation.saved,
-            inactiveTitle: Localization.Recommendation.save,
-            highlightedColor: .ui.coral1,
-            activeColor: .ui.coral2
-        ) {
-            model.saveAction(isSaved: currentSavedItem != nil && currentSavedItem?.isArchived == false)
-        }
-        .accessibilityIdentifier("save-button")
-    }
-
-    /// Overflow menu
-    func makeOverflowMenu() -> some View {
-        Menu {
-            ForEach(model.overflowActions, id: \.self) { buttonAction in
-                if let title = buttonAction.title {
-                    Button(action: {
-                        buttonAction.action()
-                    }) {
-                        Text(title)
-                    }
-                }
-            }
-        } label: {
-            Image(systemName: "ellipsis")
-                .foregroundColor(Color(.ui.saveButtonText))
-        }
-        .accessibilityIdentifier("overflow-button")
     }
 }
